Windows: do not use 'uname' for doing OS validation
The idea of the OS validator is to check if the host OS is supported.
e.g. 'uname' would return 'Linux' on Linux and something else on
OSX, BSD.

But 'uname' is a Unix tool, and only available on Windows if third
party packages are installed like MSYS.

- Split the os_validator*.go files for Unix and Windows.
- Align the fetching of OS name and kernel version for Windows
with the kubelet which uses the CurrentVersion reg key under HKLM.
- In the Windows spec use 'Windows Server' as the required name
and match it in 'CurrentVersion.ProductName' as a prefix.
- Add OS specific unit tests.